refactor: decompose kzg module into dir
