+++ src/kzg/setup.rs
@@ -4,21 +4,17 @@ use std::{
     path::Path,
 };
 
+use super::{Commitment, Polynomial, Proof};
 use crate::{
     blob::Blob,
-    bls::{self, Fr, Scalar, P1, P2},
-    math::{self, BitReversalPermutation},
+    bls::{self, ECGroupError, Error as BlsError, Fr, Scalar, P1, P2},
 };
 
 use alloy_primitives::{hex, Bytes, FixedBytes};
 
-pub enum Error {
-    Bls(bls::Error),
-}
-
 #[derive(Debug)]
 pub enum LoadSetupError {
-    Bls(bls::Error),
+    Bls(BlsError),
     Io(io::Error),
     Hex(hex::FromHexError),
     Serde(serde_json::Error),
@@ -55,28 +51,28 @@ impl<const G1: usize, const G2: usize> Setup<G1, G2> {
         let mut g1_lagrange: Box<[P1; G1]> = Box::new([P1::default(); G1]);
         for (i, point) in setup.g1_lagrange.iter().enumerate() {
             if point.len() != 48 {
-                return Err(LoadSetupError::Bls(bls::Error::from(
-                    bls::ECGroupError::InvalidEncoding,
+                return Err(LoadSetupError::Bls(BlsError::from(
+                    ECGroupError::InvalidEncoding,
                 )));
             }
             // TODO: skip unnecessary allocation
             let point = FixedBytes::<48>::from_slice(point);
             let point =
-                P1::deserialize(point).map_err(|err| LoadSetupError::Bls(bls::Error::from(err)))?;
+                P1::deserialize(point).map_err(|err| LoadSetupError::Bls(BlsError::from(err)))?;
             g1_lagrange[i] = point;
         }
 
         let mut g2_monomial: Box<[P2; G2]> = Box::new([P2::default(); G2]);
         for (i, point) in setup.g2_monomial.iter().enumerate() {
             if point.len() != 96 {
-                return Err(LoadSetupError::Bls(bls::Error::from(
-                    bls::ECGroupError::InvalidEncoding,
+                return Err(LoadSetupError::Bls(BlsError::from(
+                    ECGroupError::InvalidEncoding,
                 )));
             }
             // TODO: skip unnecessary allocation
             let point = FixedBytes::<96>::from_slice(point);
             let point =
-                P2::deserialize(point).map_err(|err| LoadSetupError::Bls(bls::Error::from(err)))?;
+                P2::deserialize(point).map_err(|err| LoadSetupError::Bls(BlsError::from(err)))?;
             g2_monomial[i] = point;
         }
 
@@ -209,120 +205,6 @@ impl<const G1: usize, const G2: usize> Setup<G1, G2> {
 impl<const G1: usize, const G2: usize> AsRef<Setup<G1, G2>> for Setup<G1, G2> {
     fn as_ref(&self) -> &Setup<G1, G2> {
         self
-    }
-}
-
-pub(crate) struct Polynomial<const N: usize>(pub(crate) Box<[Fr; N]>);
-
-impl<const N: usize> Polynomial<N> {
-    /// evaluates the polynomial at `point`.
-    pub(crate) fn evaluate(&self, point: Fr) -> Fr {
-        let roots = math::roots_of_unity::<N>();
-        let roots = BitReversalPermutation::new(roots);
-
-        // if `point` is a root of a unity, then we have the evaluation available
-        for i in 0..N {
-            if point == roots[i] {
-                return self.0[i];
-            }
-        }
-
-        let mut eval = Fr::ZERO;
-
-        // barycentric evaluation summation
-        for i in 0..N {
-            let numer = self.0[i] * roots[i];
-            let denom = point - roots[i];
-            let term = numer / denom;
-            eval = eval + term;
-        }
-
-        // barycentric evaluation scalar multiplication
-        let term = (point.pow(Fr::from(N as u64)) - Fr::ONE) / Fr::from(N as u64);
-        eval * term
-    }
-
-    /// returns a `Proof` for the evaluation of the polynomial at `point`.
-    pub(crate) fn prove<const G1: usize, const G2: usize>(
-        &self,
-        point: Fr,
-        setup: impl AsRef<Setup<G1, G2>>,
-    ) -> (Fr, Proof) {
-        assert_eq!(G1, N);
-        let roots = math::roots_of_unity::<N>();
-        let roots = BitReversalPermutation::new(roots);
-
-        let eval = self.evaluate(point);
-
-        // compute the quotient polynomial
-        //
-        // TODO: parallelize (e.g. rayon)
-        let mut quotient_poly = Vec::with_capacity(N);
-        for i in 0..N {
-            let numer = self.0[i] - eval;
-            let denom = roots[i] - point;
-            let quotient = if denom != Fr::ZERO {
-                numer / denom
-            } else {
-                let mut quotient = Fr::ZERO;
-                for j in 0..N {
-                    if j == i {
-                        continue;
-                    }
-
-                    let coefficient = self.0[j] - eval;
-                    let numer = coefficient * roots[j];
-                    let denom = (roots[i] * roots[i]) - (roots[i] * roots[j]);
-                    let term = numer / denom;
-                    quotient = quotient + term;
-                }
-                quotient
-            };
-            quotient_poly.push(Scalar::from(quotient));
-        }
-
-        let g1_lagrange = BitReversalPermutation::new(setup.as_ref().g1_lagrange.as_slice());
-        let lincomb = P1::lincomb(g1_lagrange.iter().zip(quotient_poly));
-
-        (eval, Proof(lincomb))
-    }
-}
-
-#[derive(Clone, Debug, Default, Eq, PartialEq)]
-pub struct Commitment(pub(crate) P1);
-
-impl Commitment {
-    pub const BYTES: usize = P1::BYTES;
-
-    pub fn deserialize<T: AsRef<[u8; Self::BYTES]>>(bytes: T) -> Result<Self, Error> {
-        P1::deserialize(bytes)
-            .map(Self)
-            .map_err(|err| Error::Bls(bls::Error::from(err)))
-    }
-}
-
-impl From<P1> for Commitment {
-    fn from(point: P1) -> Self {
-        Self(point)
-    }
-}
-
-#[derive(Clone, Debug, Default, Eq, PartialEq)]
-pub struct Proof(pub(crate) P1);
-
-impl Proof {
-    pub const BYTES: usize = P1::BYTES;
-
-    pub fn deserialize<T: AsRef<[u8; Self::BYTES]>>(bytes: T) -> Result<Self, Error> {
-        P1::deserialize(bytes)
-            .map(Self)
-            .map_err(|err| Error::Bls(bls::Error::from(err)))
-    }
-}
-
-impl From<P1> for Proof {
-    fn from(point: P1) -> Self {
-        Self(point)
     }
 }
